Allow to set a specific port for the tests
Previous implementation always used port 8080 for tests, which is not
ideal. This implementation allows the user to set a specific port for
each test in a specific module. This way we can test different endpoints
from different modules.

In order to set the port to use, the `port` macro is called. This macro
expands to the definition of the `__RESTEST_INTERNAL__PORT` constant,
whose type is u16. This constant is then reused in the `assert_api`
macro.