Move request-related code out of the macro
Previous request-related code was part of the macro expansion. The
problem with macros is that all the static analysis is performed when
the user uses restsest *as a dependency*, which can potentially make us
miss some important bugs.

This moves all the request-related code in a specific, hidden function.
It is reponsible for serializing the request body, creating and
performing the request and deserializing the body. All these steps are
now checked *inside* our crate, instead of being checked at each macro
call.