Add variable binding (#4)
With previous implementation, it was not possible to reuse data that has been returned by the server (for instance for next requests). This PR adds this functionality.

Under the hood, the `assert_api` macro expands to a `let ... else`, which allows to expose the pattern to the outermost scope. This makes possible to reuse values, as seen in the documentation and examples.

This functionality is only available on nightly and requires the `nightly` feature flag.