Replace method chaining with carefully crafted patterns (#8)
Previous implementation used method chaining to bind values to the scope
and the `assert_matches` function was reponsible for ensuring everything
matches. This had a lot of problems such as making harder type inference
and forcing us to re-invent pattern-match syntax.

This commit reimplements the whole process with a different approach.
Internally, we use match. This has two problems:
  1. We can not bind variables with current scope with match,
  2. We can not match Strings with string literals in patterns.

Problem 1 can be fixed by recursively visiting the pattern and keeping
track of which variables are bound to the scope, so that we can pack
them in a tuple, and destructure them in the local scope.

Problem 2 can be fixed by replacing every literal with an internal
binding, and checking for its value if an additional guard.

For instance, the following code:

```rust
assert_body_matches!(
    body,
    User {
        id,
        name: "Grace Hopper",
    },
};
```

Expands to:

```rust
let (id,) = match body {
    User {
        id,
        name: __restest_internal_0,
    } if __restest_internal_0 == "Grace Hopper" => (name,),

    _ => panic!(),
}
```